refactor: consolidate into unified node with resource groups
- Merge DocutrayKnowledgeBase functionality into main Docutray node
- Implement resource-based organization (Document/Knowledge Base)
- Add dynamic knowledge base selector with dropdown
- Create modular description files for better code organization
- Follow Slack node pattern from n8n-nodes-base
- Update documentation to reflect unified architecture
- Maintain all existing functionality with improved UX
