+++ CLAUDE.md
@@ -13,28 +13,35 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 
 ## Architecture
 
-This is an n8n community node package for Docutray integration using the n8n nodes API version 1. The package provides two specialized nodes: one for OCR document processing and document type identification, and another for knowledge base search using RAG technology. The project follows n8n's standard node development patterns:
+This is an n8n community node package for Docutray integration using the n8n nodes API version 1. The package provides a unified node with resource-based organization for OCR document processing, document type identification, and knowledge base search using RAG technology. The project follows n8n's standard node development patterns:
 
 ### Project Structure
 - `/nodes/` - Contains node implementations (*.node.ts files)
-  - `Docutray/` - Document processing node (Convert and Identify operations)
-  - `DocutrayKnowledgeBase/` - Knowledge base search node (Search operation)
+  - `Docutray/` - Unified node with resource-based organization
+    - `Docutray.node.ts` - Main node implementation with resource routing
+    - `DocutrayDescription.ts` - Main description combining all resources
+    - `DocumentDescription.ts` - Document resource operations (Convert/Identify)
+    - `KnowledgeBaseDescription.ts` - Knowledge base resource operations (Search)
 - `/credentials/` - Contains credential type definitions (*.credentials.ts files)
-  - Expected: DocutrayApi credentials for API key authentication (shared between both nodes)
+  - `DocutrayApi.credentials.ts` - API key authentication (shared across all resources)
 - `/dist/` - Compiled output (generated by build process)
 - `gulpfile.js` - Handles icon copying during build
 
 ### Node Implementation Pattern
-Nodes implement the `INodeType` interface with:
+The unified node implements the `INodeType` interface with:
 - `description: INodeTypeDescription` - Defines UI, properties, credentials, and API defaults
-- **Docutray node**: Resource/Operation pattern (Document resource with Convert/Identify operations)
-- **DocutrayKnowledgeBase node**: Direct operation pattern (Search operation only)
+- **Resource/Operation pattern**: Users first select resource (Document/Knowledge Base), then operation
+- **Resource routing in execute method**: Handles different resources with appropriate API calls
+- **LoadOptions methods**: Dynamic dropdown population for knowledge base selection
+- **Modular description files**: Separate files for each resource's operations and fields
 - TypeScript compilation to CommonJS modules targeting ES2019
 
 ### Docutray API Operations
-- **Convert Operation**: OCR text extraction from documents (supports file upload, Base64, and URL input)
-- **Identify Operation**: Document type identification with confidence scores
-- **Knowledge Base Search Operation**: RAG-based semantic search in knowledge bases
+- **Document Resource**:
+  - **Convert Operation**: OCR text extraction from documents (supports file upload, Base64, and URL input)
+  - **Identify Operation**: Document type identification with confidence scores
+- **Knowledge Base Resource**:
+  - **Search Operation**: RAG-based semantic search in knowledge bases with dynamic selector
 
 ### Credentials Pattern
 Credentials implement `ICredentialType` interface with:
@@ -58,6 +65,8 @@ Install n8n globally (`npm install n8n -g`) and link this package for local test
 - **Authentication**: Bearer token (API key)
 - **Convert Endpoint**: `/convert/post` - OCR text extraction
 - **Identify Endpoint**: `/identify/post` - Document type identification
-- **Knowledge Base Search Endpoint**: `/knowledge/bases/{id}/search/post` - RAG semantic search
+- **Knowledge Base List Endpoint**: `/knowledge-bases` - List available knowledge bases for dynamic selector
+- **Knowledge Base Search Endpoint**: `/knowledge-bases/{id}/search/post` - RAG semantic search
 - **Supported formats**: JPEG, PNG, GIF, BMP, WebP, PDF (max 100MB)
 - **Input methods**: File upload, Base64 encoding, or image URL
+- **Production endpoint**: `https://app.docutray.com/api`
